Avance para agregar notas
